+++ packages/models/src/Domain/Syncable/SmartView/SmartView.ts
@@ -6,6 +6,9 @@ import { SystemViewId } from './SystemViewId'
 import { EmojiString, IconType } from '../../Utilities/Icon/IconType'
 import { SmartViewDefaultIconName, systemViewIcon } from './SmartViewIcons'
 import { SmartViewContent } from './SmartViewContent'
+import { TagPreferences } from '../Tag/TagPreferences'
+import { ItemInterface } from '../../Abstract/Item'
+import { ContentType } from '@standardnotes/common'
 
 export const SMART_TAG_DSL_PREFIX = '!['
 
@@ -13,10 +16,13 @@ export function isSystemView(view: SmartView): boolean {
   return Object.values(SystemViewId).includes(view.uuid as SystemViewId)
 }
 
+export const isSmartView = (x: ItemInterface): x is SmartView => x.content_type === ContentType.SmartView
+
 export class SmartView extends DecryptedItem<SmartViewContent> {
   public readonly predicate!: PredicateInterface<DecryptedItem>
   public readonly title: string
   public readonly iconString: IconType | EmojiString
+  public readonly preferences?: TagPreferences
 
   constructor(payload: DecryptedPayloadInterface<SmartViewContent>) {
     super(payload)
@@ -27,6 +33,8 @@ export class SmartView extends DecryptedItem<SmartViewContent> {
     } else {
       this.iconString = this.payload.content.iconString || SmartViewDefaultIconName
     }
+
+    this.preferences = this.payload.content.preferences
 
     try {
       this.predicate = this.content.predicate && predicateFromJson(this.content.predicate)