+++ packages/services/src/Domain/Device/MobileDeviceInterface.ts
@@ -16,7 +16,7 @@ export interface MobileDeviceInterface extends DeviceInterface {
   shareBase64AsFile(base64: string, filename: string): Promise<void>
   downloadBase64AsFile(base64: string, filename: string, saveInTempLocation?: boolean): Promise<string | undefined>
   previewFile(base64: string, filename: string): Promise<boolean>
-  confirmAndExit(): void
+  exitApp(confirm?: boolean): void
   addComponentUrl(componentUuid: string, componentUrl: string): void
   removeComponentUrl(componentUuid: string): void
   isUrlComponentUrl(url: string): boolean