+++ packages/services/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.37.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.36.0...@standardnotes/services@1.37.0) (2022-10-25)
+
+### Features
+
+* per-tag display preferences ([#1868](https://github.com/standardnotes/app/issues/1868)) ([ee7f11c](https://github.com/standardnotes/app/commit/ee7f11c933285234b20fa2d5db4fcf0d6a5dce9d))
+
+# [1.36.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.35.0...@standardnotes/services@1.36.0) (2022-10-24)
+
+### Features
+
+* add preference on android to not confirm when exiting ([#1870](https://github.com/standardnotes/app/issues/1870)) ([28ef905](https://github.com/standardnotes/app/commit/28ef9054fad5236bc9f1fc87768d3c17059da4fb))
+
 # [1.35.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.34.4...@standardnotes/services@1.35.0) (2022-10-24)
 
 ### Features